Bugfix for model saving.

+++ Seq2SeqSharp/Corpus/SntBatchStreamWriter.cs
@@ -22,16 +22,20 @@ namespace Seq2SeqSharp.Corpus
         static object locker = new object();
         private List<string> outputBuffer = new List<string>();
         private List<string> alignmentBuffer = new List<string>();
-        StreamWriter sw;
+        StreamWriter sw = null;
         StreamWriter swAlignment = null;
         int endWriteIdx = 0;
         int numWrittenToFile = 0;
 
         SentencePiece sp;
 
-        public SntBatchStreamWriter(string filePath, string sentencePieceModelPath = null, string alignmentFilePath = null)
+        public SntBatchStreamWriter(string filePath = null, string sentencePieceModelPath = null, string alignmentFilePath = null)
         {
-            sw = new StreamWriter(filePath, false);
+            if (String.IsNullOrEmpty(filePath) == false)
+            {
+                sw = new StreamWriter(filePath, false);
+            }
+
             if (String.IsNullOrEmpty(alignmentFilePath) == false)
             {
                 swAlignment = new StreamWriter(alignmentFilePath, false);
@@ -49,45 +53,47 @@ namespace Seq2SeqSharp.Corpus
         {
             lock (locker)
             {
-                List<string> finalOutputs = new List<string>();
-
-                List<List<string>> outputAllTasks = new List<List<string>>();
-                foreach (NetworkResult result in results) //for result in each task, shape [beam size, batch size, tgt token size]
+                List<string> finalOutputs = null;
+                int batchSize = results[0].Output[0].Count;
+                if (sw != null)
                 {
-                    List<string> outputEntireBatch = new List<string>();
-                    for (int batchIdx = 0; batchIdx < result.Output[0].Count; batchIdx++)
+                    finalOutputs = new List<string>();
+                    List<List<string>> outputAllTasks = new List<List<string>>();
+                    foreach (NetworkResult result in results) //for result in each task, shape [beam size, batch size, tgt token size]
                     {
-                        List<string> outputBeams = new List<string>();
-                        for (int beamIdx = 0; beamIdx < result.Output.Count; beamIdx++)
+                        List<string> outputEntireBatch = new List<string>();
+                        for (int batchIdx = 0; batchIdx < result.Output[0].Count; batchIdx++)
                         {
-                            string output = string.Join(" ", result.Output[beamIdx][batchIdx]);
-                            outputBeams.Add(output);
+                            List<string> outputBeams = new List<string>();
+                            for (int beamIdx = 0; beamIdx < result.Output.Count; beamIdx++)
+                            {
+                                string output = string.Join(" ", result.Output[beamIdx][batchIdx]);
+                                outputBeams.Add(output);
+                            }
+
+                            string outputAllBeams = string.Join("\t", outputBeams);
+                            outputEntireBatch.Add(outputAllBeams);
                         }
 
-                        string outputAllBeams = string.Join("\t", outputBeams);
-                        outputEntireBatch.Add(outputAllBeams);
+                        outputAllTasks.Add(outputEntireBatch);
                     }
 
-                    outputAllTasks.Add(outputEntireBatch);
-                }
-
-                int batchSize = outputAllTasks[0].Count;
-                for (int i = 0; i < batchSize; i++) // iterate all output line
-                {
-                    List<string> output = new List<string>();
-                    for (int j = 0; j < outputAllTasks.Count; j++) // iterate all tasks result
+                    for (int i = 0; i < batchSize; i++) // iterate all output line
                     {
-                        output.Add(outputAllTasks[j][i]);
-                    }
+                        List<string> output = new List<string>();
+                        for (int j = 0; j < outputAllTasks.Count; j++) // iterate all tasks result
+                        {
+                            output.Add(outputAllTasks[j][i]);
+                        }
 
-                    finalOutputs.Add(string.Join("\t", output));
+                        finalOutputs.Add(string.Join("\t", output));
+                    }
                 }
 
                 List<string> finalAlignments = null;
                 if (swAlignment != null)
                 {
                     finalAlignments = new List<string>();
-
                     List<List<string>> alignmentAllTasks = new List<List<string>>();
                     foreach (NetworkResult result in results) //for result in each task, shape [beam size, batch size, tgt token size]
                     {
@@ -125,46 +131,62 @@ namespace Seq2SeqSharp.Corpus
                     }
                 }
 
-                //Need to extend output buffer
-                while (outputBuffer.Count < idx + batchSize - numWrittenToFile)
+                int buffSize = 0;
+                if (sw != null)
                 {
-                    outputBuffer.Add(null);
+                    //Need to extend output buffer
+                    while (outputBuffer.Count < idx + batchSize - numWrittenToFile)
+                    {
+                        outputBuffer.Add(null);
+                    }
 
-                    if (swAlignment != null)
+                    for (int i = 0; i < batchSize; i++)
+                    {
+                        outputBuffer[idx + i - numWrittenToFile] = finalOutputs[i];
+                    }
+                    buffSize = outputBuffer.Count;
+                }
+
+                if (swAlignment != null)
+                {
+                    //Need to extend output buffer
+                    while (alignmentBuffer.Count < idx + batchSize - numWrittenToFile)
                     {
                         alignmentBuffer.Add(null);
                     }
-                }
 
-                for (int i = 0; i < batchSize; i++)
-                {
-                    outputBuffer[idx + i - numWrittenToFile] = finalOutputs[i];
-
-                    if (swAlignment != null)
+                    for (int i = 0; i < batchSize; i++)
                     {
                         alignmentBuffer[idx + i - numWrittenToFile] = finalAlignments[i];
                     }
+                    buffSize = alignmentBuffer.Count;
                 }
 
 
-
-                while(endWriteIdx < outputBuffer.Count)
+                while (endWriteIdx < buffSize)
                 {
-                    if (outputBuffer[endWriteIdx] == null)
+                    if (sw != null)
                     {
-                        break;
-                    }
+                        if (outputBuffer[endWriteIdx] == null)
+                        {
+                            break;
+                        }
 
-                    string line = outputBuffer[endWriteIdx];
-                    if (sp != null)
-                    {
-                        line = sp.Decode(line);
+                        string line = outputBuffer[endWriteIdx];
+                        if (sp != null)
+                        {
+                            line = sp.Decode(line);
+                        }
+                        sw.WriteLine(line);
                     }
-
-                    sw.WriteLine(line);
 
                     if (swAlignment != null)
                     {
+                        if (alignmentBuffer[endWriteIdx] == null)
+                        {
+                            break;
+                        }
+
                         swAlignment.WriteLine(alignmentBuffer[endWriteIdx]);
                     }
 
@@ -173,7 +195,11 @@ namespace Seq2SeqSharp.Corpus
 
                 if (endWriteIdx > 0)
                 {
-                    outputBuffer.RemoveRange(0, endWriteIdx);
+                    if (sw != null)
+                    {
+                        outputBuffer.RemoveRange(0, endWriteIdx);
+                    }
+
                     if (swAlignment != null)
                     {
                         alignmentBuffer.RemoveRange(0, endWriteIdx);
@@ -188,7 +214,10 @@ namespace Seq2SeqSharp.Corpus
 
         public void Close()
         {
-            sw.Close();
+            if (sw != null)
+            {
+                sw.Close();
+            }
 
             if (swAlignment != null)
             {
